rules: remove FromHasPublicStatusIfFromIsTableAndToIsRowLevelTTL
This is only necessary in 22.2 for compatibility with 22.1.
The clause has therefore been moved to release_22_2.

Release note: None
